requirements.txt
streamlit==1.28.1
pandas==2.1.3
numpy>=1.26.0
scikit-learn==1.3.2
statsmodels==0.14.0
scipy==1.11.4
plotly==5.17.0
seaborn==0.13.0
matplotlib==3.8.2
openpyxl==3.1.2
xlrd==2.0.1
sqlalchemy==2.0.23
streamlit-authenticator==0.2.3
stripe==7.8.0
openai==1.3.7
python-dotenv==1.0.0
bcrypt==4.1.2
streamlit-option-menu==0.3.6
geopandas==0.14.1
pycountry==23.12.11
folium==0.15.0
streamlit-folium==0.15.1
